requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-APScheduler==1.13.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.0.3
openpyxl==3.1.2
gevent==23.9.1
psutil==5.9.8
chardet>=5.0.0
py-ip2region
